Semaphore tokens can be acquired in an interrupt context.
The function osSemaphoreWait() can be called in an interrupt context given that the timeout
is set to 0ms. If a token is available, it will be acquired and the number of tokens is returned.
Otherwise, the return value is 0.

Now there is a strong symmetry between osMessagePut() <-> osSemaphoreRelease() and
osMessageGet() <-> osSemaphoreWait(): Putting a message into the queue is similar to giving a token
back to the semaphore and fetching a message from the queue is like acquiring a semaphore token.

This addition allows to couple a memory pool with a semaphore to create a blocking memory pool,
which can be used in a thread and an interrupt context.

    typedef struct
    {
        osSemaphoreId sem;
        osPoolId memPool;
    } BlockingPool;

    // Allocate a pool element. Can be called from an ISR if millisecs == 0.
    void* BlockingPool_allocate(BlockingPool* pool, uint32_t millisecs)
    {
        if (osSemaphoreWait(pool->sem, millisecs) >= 0)
            return osPoolAlloc(pool->memPool);  // This always succeeds.
    }

    // Free a pool element. Can be called from an ISR.
    void BlockingPool_free(BlockingPool* pool, void* mem)
    {
        osPoolFree(pool->memPool, mem);
        osSemaphoreRelease(pool->sem);
    }