Add a non-recursive mutex.
As recursive mutexes are considered to be less secure compared to plain mutexes, let's add a
non-recursive variant to the RTOS. A padding byte of the mutex structure can been used for
storing a flag which discriminates recursive and non-recursive mutexes. If a task tries
to re-lock a plain mutex which it has already acquired, an error is returned.

In order not to break the existing API, a function osPlainMutexCreate() has been added to
create a plain (non-recursive) mutex.