include/cmsis_gcc.h: only read define if it exists
If CMSIS is used on a cortex-m core,
__CORTEX_M will be defined to be the correct
core version, while __CORTEX_SC will be undefined.

If CMSIS is used on a ARM SecureCore core,
__CORTEX_SC will be defined to be the correct
core version, while __CORTEX_M will be undefined.

Thus, only one of the two defines will
exists at a time.

In order to avoid compiler warnings, this
patch adds checks for the existence of the
defines before accessing them.